fix: support importing project without model

+++ api/routes/port_project.js
@@ -52,9 +52,10 @@ const nativizeProject = function (projectId, projectName, backup) {
         if (!Object.keys(allCollections).includes(col)) delete nativizedBackup[col];
     });
 
-    let nlu_models = project.nlu_models;
+    let nlu_models = undefined;
 
     if ('models' in nativizedBackup) {
+        nlu_models = project.nlu_models;
         const modelMapping = {};
         nativizedBackup.models.forEach((m) =>
             Object.assign(modelMapping, { [m._id]: uuidv4() }),
@@ -348,9 +349,10 @@ exports.importProject = async function (req, res) {
             await overwriteCollection(projectId, project.nlu_models, col, backup);
         }
         const overwrittenProject = { ...project, ...backup.project };
-        await NLUModels.deleteMany({ _id: { $in: project.nlu_models } }).exec();
+        if ('models' in backup) await NLUModels.deleteMany({ _id: { $in: project.nlu_models } }).exec();
+        else overwrittenProject.nlu_models = project.nlu_models;
         await Projects.deleteMany({ _id: projectId }).exec();
-        await NLUModels.insertMany(backup.models);
+        if ('models' in backup) await NLUModels.insertMany(backup.models);
         await Projects.insertMany([overwrittenProject]);
         return res.status(200).send('Success');
     } catch (e) {
